% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_scenario.R
\name{as_tibble.tidyrisk_scenario}
\alias{as_tibble.tidyrisk_scenario}
\alias{as.data.frame.tidyrisk_scenario}
\title{Coerce the parameters of a tidyrisk_scenario to a tibble}
\usage{
\method{as_tibble}{tidyrisk_scenario}(x, ...)

\method{as.data.frame}{tidyrisk_scenario}(x, ...)
}
\arguments{
\item{x}{A tidyrisk_scenario}

\item{...}{Currently not used}
}
\description{
Coerce the parameters of a tidyrisk_scenario to a tibble
}
