% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_domain_impact}
\alias{calculate_domain_impact}
\title{Calculate quantified impact at a domain level}
\usage{
calculate_domain_impact(domain_summary, domains)
}
\arguments{
\item{domain_summary}{Domain-level summary of \code{run_simulation} results.}

\item{domains}{Dataframe of all domains in scope.}
}
\value{
A dataframe
}
\description{
Calculate quantified impact at a domain level
}
