% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_weak_domains}
\alias{calculate_weak_domains}
\title{Calculate control weaknesses on a domain level.}
\usage{
calculate_weak_domains(simulation_results, domains)
}
\arguments{
\item{simulation_results}{Results of running the risk simulations.}

\item{domains}{Domain titles and IDs as a dataframe.}
}
\value{
Control weaknesses summarized by domain.
}
\description{
Calculate control weaknesses on a domain level.
}
