% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capabilities}
\alias{capabilities}
\title{Capabilities}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 4 columns.}
\usage{
capabilities
}
\description{
A qualitative dataset of sample capabilities and qualitative level of effectiveness.
}
\details{
\describe{
\item{id}{unique id of the capability}
\item{domain_id}{domain id to which the capability applies}
\item{capability}{full text summary of the capability}
\item{diff}{qualitative label of control effectiveness}
}
}
\keyword{datasets}
