% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_availability}
\alias{check_availability}
\title{Helper function to verify package availability}
\usage{
check_availability(packages, func)
}
\arguments{
\item{packages}{Packages to verify are available}

\item{func}{Calling function}
}
\value{
Invisible
}
\description{
Helper function to verify package availability
}
\examples{
\dontrun{
check_availability(packages = c("ggplot2", "dplyr"), func = "my_function")
}
}
