% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{create_templates}
\alias{create_templates}
\title{Create initial template files}
\usage{
create_templates(base_directory = "~/evaluator")
}
\arguments{
\item{base_directory}{Parent directory under which to create starter files.}
}
\value{
A dataframe of the starter filenames, along with a flag on whether a file was copied.
}
\description{
Given a base directory, copy the provided sample files into an \code{inputs}
subdirectory. This makes the starter files available for customaizing and
data collection. The \code{inputs} directory will be created if not already present.
Pre-existing files, if present, will not be overwritten.
}
\examples{
\dontrun{
create_templates("~/evaluator")
}
}
