% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidyrisk_skeleton.R
\name{create_tidyrisk_scenario_skeleton}
\alias{create_tidyrisk_scenario_skeleton}
\title{Create a skeleton tidyrisk scenario object in the current document}
\usage{
create_tidyrisk_scenario_skeleton()
}
\description{
Inserts a code block into the active document in RStudio for creating a
\link{tidyrisk_scenario} object. This is an easy way of rapidly running
a simulation.
}
