% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{derive_control_key}
\alias{derive_control_key}
\title{Derive control ID to control description mappings}
\usage{
derive_control_key(capability_ids, capabilities)
}
\arguments{
\item{capability_ids}{Comma-delimited list of capabilities in scope for a scenario.}

\item{capabilities}{Dataframe of master list of all qualitative capabilities.}
}
\value{
A named list of control IDs and descriptions.
}
\description{
Given a comma-separated list of control IDs, return a named list
of descriptions for each control with the names set to the control
IDs.
}
\examples{
data(capabilities)
capability_ids <- c("1, 3")
derive_control_key(capability_ids, capabilities)
}
