% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{detail_vuln}
\alias{detail_vuln}
\title{Calculate the vulnerability}
\usage{
detail_vuln(tc, diff)
}
\arguments{
\item{tc}{Threat capability (as a percentage)}

\item{diff}{Difficult (as a percentage)}
}
\value{
List containing samples (as a vector) and details (as a list)
}
\description{
Composition function for use in sample_vuln, does a simple compare of
all occurances where the threat capability (TC) is greater than the
difficulty (DIFF).
}
