% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain_summary}
\alias{domain_summary}
\title{Domain-level risk summary}
\format{\describe{
\item{domain_id}{abbreviated name of the domain}
\item{loss_events_mean}{mean number of loss events}
\item{loss_events_min}{minimum number of loss events}
\item{loss_events_max}{maximum number of loss events}
\item{loss_events_median}{median number of loss events}
\item{ale_max}{minimum annual loss expected}
\item{ale_median}{median annual loss expected}
\item{ale_mean}{mean annual loss expected}
\item{ale_max}{maximum annual loss expected}
\item{ale_sd}{standard deviation annual loss expected}
\item{ale_var}{value at risk, ale}
\item{mean_threat_events}{mean threat events}
\item{mean_avoided_events}{mean avoided events}
\item{mean_tc_exceedance}{mean threat capability exceedance}
\item{mean_diff_exceedance}{mean difficulty exceedance}
\item{mean_vuln}{mean vulnerability of the scenario}
}}
\usage{
domain_summary
}
\description{
A dataset of quantified information security risk, summarized at the
domain level.
}
\keyword{datasets}
