% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domains}
\alias{domains}
\title{Domain mappings}
\format{\describe{
\item{domain_id}{abbreviated name of the domain}
\item{domain}{full title of the domain}
}}
\usage{
domains
}
\description{
A dataset of domains and domain IDs.
}
\keyword{datasets}
