% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator-package.R
\docType{package}
\name{evaluator-package}
\alias{evaluator}
\alias{evaluator-package}
\title{evaluator: Quantified Risk Assessment Toolkit}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An open source risk analysis toolkit based on the OpenFAIR ontology 
  <https://publications.opengroup.org/c20b> and risk analysis standard 
  <https://publications.opengroup.org/c20a>. Empowers an organization to 
  perform a quantifiable, repeatable, and data-driven risk review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://evaluator.tidyrisk.org}
  \item Report bugs at \url{https://github.com/davidski/evaluator/issues}
}

}
\author{
\strong{Maintainer}: David Severski \email{davidski@deadheaven.com} (\href{https://orcid.org/0000-0001-7867-0459}{ORCID})

}
\keyword{internal}
