% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_factor.R
\name{evaluator_elem}
\alias{evaluator_elem}
\alias{new_evaluator_elem}
\title{Construct a risk element object}
\usage{
new_evaluator_elem(samples = vector(), factor_label = character(),
  details = list())

evaluator_elem(samples, factor_label, details = list())
}
\arguments{
\item{samples}{samples}

\item{factor_label}{fl}

\item{details}{details}
}
\description{
Construct a risk element object
}
