% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{explore_scenarios}
\alias{explore_scenarios}
\title{Launch the Scenario Explorer web application.}
\usage{
explore_scenarios(input_directory = "data", results_directory = "results")
}
\arguments{
\item{input_directory}{Location of input files}

\item{results_directory}{Location of simulation results}
}
\value{
Invisible NULL
}
\description{
Launch the Scenario Explorer web application.
}
