% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{generate_event_outcomes_plot}
\alias{generate_event_outcomes_plot}
\title{Display a scatterplot for a particular scenario ID.}
\usage{
generate_event_outcomes_plot(control_weakness)
}
\arguments{
\item{control_weakness}{Domain-level control weakness from \code{calculate_control_weakness}}
}
\value{
ggplot object
}
\description{
Display a scatterplot for a particular scenario ID.
}
