% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\title{Generate sample analysis report}
\usage{
generate_report(input_directory = "~/data", results_directory = "~/results",
  output_file, styles = NULL, focus_scenario_ids = c(51, 12),
  format = "html", ...)
}
\arguments{
\item{input_directory}{Location of input files.}

\item{results_directory}{Location of simulation results.}

\item{output_file}{Full path to output file.}

\item{styles}{Optional full path to CSS file to overide default styles.}

\item{focus_scenario_ids}{IDs of scenarios of special interest.}

\item{format}{Format to generate (html, pdf, word).}

\item{...}{Any other parameters to pass straight to \code{rmarkdown::html_document}.}
}
\value{
Default return values of the \code{rmarkdown::render} function.
}
\description{
Generate sample analysis report
}
\examples{
\dontrun{
generate_report("~/inputs", "~/results", "~/risk_report.html")
}
}
