% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{generate_scatterplot-deprecated}
\alias{generate_scatterplot-deprecated}
\title{Display a scatterplot for a particular scenario ID}
\arguments{
\item{simulation_results}{Simulation results from \code{run_simulations}.}

\item{scenario_id}{ID of the scenario to display.}
}
\value{
A ggplot object.
}
\description{
Given a detailed results dataframe and a specific scenario identifier,
create a scatterplot of the number of loss events versus the total amount of
expected annual loss expected for each simulation. This provides a
detailed view on the results for a particular scenario.
}
\examples{
\dontrun{
data(mc_simulation_results)
generate_scatterplot(mc_simulation_results, scenario_id = "RS-50")
}
}
\seealso{
\code{\link{evaluator-deprecated}}

Other result graphs: \code{\link{exposure_histogram}},
  \code{\link{generate_event_outcomes_plot}},
  \code{\link{generate_heatmap}},
  \code{\link{loss_exceedance_curve}},
  \code{\link{loss_scatterplot}}
}
\concept{result graphs}
\keyword{internal}
