% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{get_base_fontfamily}
\alias{get_base_fontfamily}
\title{Select a base graphics font family}
\usage{
get_base_fontfamily()
}
\value{
String of the preferred base font
}
\description{
The Benton Sans Regular font is preferred. When Benton is not available
via \code{extrafont}, Evaluator uses a fallback of Arial Narrow.
}
