% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_spreadsheet}
\alias{import_spreadsheet}
\title{Import the scenario spreadsheet}
\usage{
import_spreadsheet(
  survey_file = system.file("survey", "survey.xlsx", package = "evaluator"),
  domains = NULL,
  output_dir = "~/evaluator/results"
)
}
\arguments{
\item{survey_file}{Path to survey Excel file. Defaults to an Evaluator-provided sample spreadsheet.}

\item{domains}{Dataframe of domains and domain IDs. Defaults to built-in sample \code{domains} dataset.}

\item{output_dir}{Output file directory.}
}
\value{
Dataframe of file information on the two newly created files.
}
\description{
This is a convenience wrapper around the \code{\link{import_scenarios}} and
\code{\link{import_capabilities}} functions. Writes cleaned
comma-separated formatted files for the scenarios and
capabilities to disk.
}
