% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_scenario.R
\name{is.tidyrisk_scenario}
\alias{is.tidyrisk_scenario}
\title{Alias test for tidyrisk scenario}
\usage{
is.tidyrisk_scenario(x)
}
\arguments{
\item{x}{An object}
}
\description{
Please use \code{is_tidyrisk_scenario()} instead.
}
