% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_scenario.R
\name{is_tidyrisk_scenario}
\alias{is_tidyrisk_scenario}
\title{Test if the object is a tidyrisk_scenario}
\usage{
is_tidyrisk_scenario(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{tidyrisk_scenario} class.
}
\description{
This function returns \code{TRUE} for tidyrisk_scenario (or subclasses)
and \code{FALSE} for all other objects.
}
