% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data-deprecated}
\alias{load_data-deprecated}
\title{Load input and results files}
\arguments{
\item{input_directory}{Location of input files.}

\item{results_directory}{Location of simulation results.}
}
\value{
List of all key data objects.
}
\description{
Given an input directory and a directory of simulation results, load all
of the key Evaluator data objects into memory.
}
\examples{
\dontrun{
load_data("~/evaluator/inputs", "~/evaluator/results")
}

}
\seealso{
\code{\link{evaluator-deprecated}}
}
\keyword{internal}
