% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load input and results files}
\usage{
load_data(input_directory = "~/data", results_directory = "~/data")
}
\arguments{
\item{input_directory}{Location of input files.}

\item{results_directory}{Location of simulation results.}
}
\value{
List of all key data objects.
}
\description{
Given a input directory and a directory of simulation results, load all
of the key Evaluator data objects into memory.
}
\examples{
\dontrun{
load_data("~/input_files", "~/simulation_results")
}
}
