% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{loss_exceedance_curve}
\alias{loss_exceedance_curve}
\title{Display the loss exceedance curve for a group of one or more scenarios}
\usage{
loss_exceedance_curve(iteration_results)
}
\arguments{
\item{iteration_results}{Iteration-level summary from \code{summarize_iterations}.}
}
\value{
A ggplot object.
}
\description{
Display the loss exceedance curve for a group of one or more scenarios
}
\examples{
data(mc_simulation_results)
summarize_iterations(mc_simulation_results$results) \%>\% loss_exceedance_curve()
}
\seealso{
Other result graphs: 
\code{\link{exposure_histogram}()},
\code{\link{generate_event_outcomes_plot}()},
\code{\link{generate_heatmap}()},
\code{\link{generate_scatterplot-deprecated}},
\code{\link{loss_scatterplot}()}
}
\concept{result graphs}
