% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mappings}
\alias{mappings}
\title{Qualitative to Quantitative Mappings}
\format{An object of class \code{data.frame} with 14 rows and 6 columns.}
\usage{
mappings
}
\description{
A dataset of sample qualitative to quantitative mappings.
}
\details{
\describe{
\item{type}{OpenFAIR taxonomy to which this mapping applies}
\item{label}{Qualitative label}
\item{l}{BetaPERT low value}
\item{ml}{BetaPERT most likely value}
\item{h}{BetaPERT high value}
\item{conf}{BetaPERT confidence value}
}
}
\keyword{datasets}
