% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_capabilities}
\alias{mc_capabilities}
\title{Capabilities}
\format{\describe{
\item{capability_id}{unique id of the capability}
\item{domain_id}{domain id to which the capability applies}
\item{capability}{full text summary of the capability}
\item{diff}{qualitative label of control effectiveness}
}}
\source{
This is hypothetical information. Any similarity to any other
entitiy is completely coincidental.
}
\usage{
mc_capabilities
}
\description{
A sample set of capabilities for the demonstration (and artificial)
MetroCare information security program.
}
\keyword{datasets}
