% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_domains}
\alias{mc_domains}
\title{Domain mappings}
\format{\describe{
\item{domain_id}{abbreviated name of the domain}
\item{domain}{full title of the domain}
}}
\source{
This is hypothetical information. Any similarity to any other
entity is completely coincidental.
}
\usage{
mc_domains
}
\description{
A sample set of the domains for the demonstration (and artificial)
MetroCare information security program.
}
\keyword{datasets}
