% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_mappings}
\alias{mc_mappings}
\title{Qualitative to quantitative mappings}
\format{\describe{
\item{type}{The element in the OpenFAIR ontology to which this mapping applies}
\item{label}{Qualitative label}
\item{l}{BetaPERT low value}
\item{ml}{BetaPERT most likely value}
\item{h}{BetaPERT high value}
\item{conf}{BetaPERT confidence value}
}}
\source{
This is hypothetical information. Any similarity to any other
entitiy is completely coincidental.
}
\usage{
mc_mappings
}
\description{
A sample set of qualitative to quantitative mappings for the demonstration
(and artificial) MetroCare information security program.
}
\keyword{datasets}
