% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_qualitative_scenarios}
\alias{mc_qualitative_scenarios}
\title{Qualitative information security risk scenarios}
\format{
\describe{
\item{scenario_id}{id of the scenario, primary key}
\item{scenario}{full text description of the risk scenario}
\item{tcomm}{full text name of threat community}
\item{tef}{qualitative label of threat frequency}
\item{tc}{qualitative label of threat capability}
\item{lm}{qualitative label of loss magnitude}
\item{domain_id}{domain id}
\item{controls}{comma delimited list of controls ids}
}
}
\source{
This is hypothetical information. Any similarity to any other
entity is completely coincidental.
}
\usage{
mc_qualitative_scenarios
}
\description{
A sample set of qualitative information security risk scenarios for the
demonstration (and artificial) MetroCare information security program.
}
\details{
No connection with any other similarly named entity is intended or implied.
}
\keyword{datasets}
