% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_scenario_summary}
\alias{mc_scenario_summary}
\title{Scenario-level risk summary}
\format{\describe{
\item{scenario_id}{ID of the scenario}
\item{domain_id}{domain id}
\item{control_description}{control description}
\item{results}{nested data frame of simulation results for the scenario}
\item{loss_events_mean}{mean number of loss events}
\item{loss_events_median}{median number of loss events}
\item{loss_events_min}{minimum number of loss events}
\item{loss_events_max}{maximum number of loss events}
\item{ale_median}{median annual loss expected}
\item{ale_max}{maximum annual loss expected}
\item{ale_var}{value at risk, ale}
\item{sle_min}{minimum single loss expectancy}
\item{sle_max}{maximum single loss expectancy}
\item{sle_mean}{mean single loss expectancy}
\item{sle_median}{median single loss expectancy}
\item{mean_tc_exceedance}{mean threat capability exceedance}
\item{mean_diff_exceedance}{mean difficulty exceedance}
\item{mean_vuln}{mean vulnerability of the scenario}
}}
\source{
This is hypothetical information. Any similarity to any other
entitiy is completely coincidental.
}
\usage{
mc_scenario_summary
}
\description{
A sample set of quantified information security risk exposure, summarized
at the scenario level, for the demonstration (and artificial)
MetroCare information security program.
}
\keyword{datasets}
