% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_simulation_results}
\alias{mc_simulation_results}
\title{Simulation results}
\format{
\describe{
\item{scenario_id}{id of the scenario}
\item{domain_id}{domain id}
\item{results}{nested data frame of simulation results for the scenario}
}
}
\source{
This is hypothetical information. Any similarity to any other
entity is completely coincidental.
}
\usage{
mc_simulation_results
}
\description{
A sample set of information security risk scenario simulation results
for the demonstration (and artificial) MetroCare information security
program.
}
\keyword{datasets}
