% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{new_evaluator_scen}
\alias{new_evaluator_scen}
\alias{evaluator_scen}
\title{Construct a quantitative scenario object}
\usage{
new_evaluator_scen(tef_params = list(), tc_params = list(),
  diff_params = list(), lm_params = list(),
  model = "openfair_tef_tc_diff_lm")

evaluator_scen(tef_params = list(), tc_params = list(),
  diff_params = list(), lm_params = list(),
  model = "openfair_tef_tc_diff_lm")
}
\arguments{
\item{tef_params}{List of threat frequency parameters}

\item{tc_params}{List of threat capability parameters}

\item{diff_params}{List of difficulty parameters}

\item{lm_params}{List of loss magnitude parameters}

\item{model}{Name of model to run}
}
\description{
Construct a quantitative scenario object
}
