% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{openfair_example}
\alias{openfair_example}
\title{Launch OpenFAIR demonstration web application}
\usage{
openfair_example()
}
\value{
Invisible NULL
}
\description{
A simple web application to try out OpenFAIR modelling.
}
\examples{
\dontrun{
openfair_example()
}
}
