% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{openfair_tef_tc_diff_lm}
\alias{openfair_tef_tc_diff_lm}
\title{Run an OpenFAIR simulation}
\usage{
openfair_tef_tc_diff_lm(scenario, diff_estimates = NULL, n = 10^4,
  title = "Untitled", verbose = FALSE)
}
\arguments{
\item{scenario}{List of tef_, tc_, and LM_ l/ml/h/conf parameters.}

\item{diff_estimates}{Parameters for estimating the scenario difficulty.}

\item{n}{Number of simulations to run.}

\item{title}{Optional name of scenario.}

\item{verbose}{Whether to print progress indicators.}
}
\value{
Dataframe of scenario name, threat_event count, loss_event count,
mean TC and DIFF exceedance, and ALE samples.
}
\description{
Run an OpenFAIR simulation
}
