% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_scenario.R
\name{print.tidyrisk_scenario}
\alias{print.tidyrisk_scenario}
\title{Default printing of a tidyrisk_scenario}
\usage{
\method{print}{tidyrisk_scenario}(x, ...)
}
\arguments{
\item{x}{A tidyrisk_scenario}

\item{...}{Currently not used}
}
\description{
Basic printing of a tidyrisk scenario
}
