% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{risk_dashboard}
\alias{risk_dashboard}
\title{Launch a single page summary risk dashboard.}
\usage{
risk_dashboard(input_directory = "~/data", results_directory = "~/results",
  ...)
}
\arguments{
\item{input_directory}{Location of input files}

\item{results_directory}{Location of simulation results}

\item{...}{Any other parameters to pass straight to \code{rmarkdown::render}}
}
\value{
Default return values of the \code{rmarkdown::render} function.
}
\description{
Launch a single page summary risk dashboard.
}
