% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_simulations}
\alias{run_simulations}
\title{Run simulations for all scenarios}
\usage{
run_simulations(scenario, simulation_count = 10000L,
  model = "openfair_tef_tc_diff_lm")
}
\arguments{
\item{scenario}{Quantitative scenarios.}

\item{simulation_count}{Number of simulations for each scenario.}

\item{model}{OpenFAIR model to use.}
}
\value{
Dataframe of raw results.
}
\description{
Run simulations for all scenarios
}
