% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_lef}
\alias{sample_lef}
\title{Sample loss event frequency}
\usage{
sample_lef(func = NULL, params = NULL)
}
\arguments{
\item{func}{Function to use to simulate LEF, defaults to `rnorm``}

\item{params}{Optional parameters to pass to `func`}
}
\value{
List containing type ("diff"), samples (as a vector), and details (as a list).
}
\description{
Sample loss event frequency
}
