% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_lm}
\alias{sample_lm}
\title{Given a number of loss events and a loss distribution, calculate losses}
\usage{
sample_lm(n, l, ml, h, conf)
}
\arguments{
\item{n}{Numer of loss events to evaluate}

\item{l}{Low boundary}

\item{ml}{Most likely}

\item{h}{High boundary}

\item{conf}{Confidence}
}
\value{
List of total loss, min/max/mean/median of sle
}
\description{
Given a number of loss events and a loss distribution, calculate losses
}
