% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_tc}
\alias{sample_tc}
\title{Sample threat capabilities (TC) from a distribution function.}
\usage{
sample_tc(n = 1, func = NULL, params = NULL)
}
\arguments{
\item{n}{Number of periods to simulate. Defaults to 1.}

\item{func}{Function to use to simulate TC, defaults to mc2d::rpert}

\item{params}{Optional parameters to pass to `func`}
}
\value{
List containing type ("t"), samples (as a list), and details (as a list).
}
\description{
Sample threat capabilities (TC) from a distribution function.
}
