% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_tef}
\alias{sample_tef}
\title{Calculate the number of simulated threat event frequencies (TEF)}
\usage{
sample_tef(func = NULL, params = NULL)
}
\arguments{
\item{func}{Function to use to simulate TEF, defaults to \code{\link[mc2d]{rpert}}.}

\item{params}{Optional parameters to pass to \code{func}.}
}
\value{
List containing type ("tef"), samples (as a vector), and details (as a list).
}
\description{
Calculate the number of simulated threat event frequencies (TEF)
}
\seealso{
Other OpenFAIR helpers: \code{\link{compare_tef_vuln}},
  \code{\link{get_mean_control_strength}},
  \code{\link{openfair_tef_tc_diff_lm}},
  \code{\link{sample_diff}}, \code{\link{sample_lef}},
  \code{\link{sample_lm}}, \code{\link{sample_tc}},
  \code{\link{sample_vuln}},
  \code{\link{select_loss_opportunities}}
}
\concept{OpenFAIR helpers}
