% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_summary}
\alias{scenario_summary}
\title{Scenario-level risk summary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56 rows and 4 columns.}
\usage{
scenario_summary
}
\description{
A dataset of quantified information security risk, summarized at the
scenario level.
}
\details{
\describe{
\item{scenario_id}{ID of the scenario}
\item{domain_id}{domain id}
\item{results}{nested data frame of simulation results for the scenario}
\item{summary}{nested data frame of summary statistics for the scenario}
}
}
\keyword{datasets}
