% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{select_events}
\alias{select_events}
\title{Calculate the threat capability and whether the control(s) resist the attack}
\usage{
select_events(n, TCestimate, DIFFsamples = NULL, DIFFestimate = NULL,
  verbose = FALSE)
}
\arguments{
\item{n}{Number of threat events to evaluate}

\item{TCestimate}{Threat capability estimate - df(l, m, h, conf)}

\item{DIFFsamples}{Pre-sampled}

\item{DIFFestimate}{Control difficulty estimate - df(l, m, h, conf)}

\item{verbose}{A boolean}
}
\value{
List of number of successful attacks (i.e. loss events),
  mean tc exceedance (how much TC > DIFF), and mean diff exceedance
  (how much DIFF > TC)
}
\description{
Calculate the threat capability and whether the control(s) resist the attack
}
