% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{select_loss_opportunities}
\alias{select_loss_opportunities}
\title{Determine which threat events result in loss opportunities}
\usage{
select_loss_opportunities(tc, diff)
}
\arguments{
\item{tc}{Threat capability (as a percentage).}

\item{diff}{Difficulty (as a percentage).}
}
\value{
List containing boolean values of length TC (as a vector) and details (as a list).
}
\description{
Composition function for use in sample_vuln, does a simple compare of
all threat events where the threat capability (TC) is greater than the
difficulty (DIFF).
}
