% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_results}
\alias{simulation_results}
\title{Information security risk simulation results}
\format{\describe{
\item{scenario_id}{id of the scenario}
\item{domain_id}{domain id}
\item{results}{nested data frame of simulation results for the scenario}
}}
\usage{
simulation_results
}
\description{
A dataset containing the full results of sample Monte Carlo
simulations of information security risk scenarios.
}
\keyword{datasets}
