% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{split_sheet}
\alias{split_sheet}
\title{Split a sheet of the survey spreadsheet into capabilities or threats}
\usage{
split_sheet(dat, table_type = "capabilities")
}
\arguments{
\item{dat}{Raw sheet input from \code{link[readxl]{readxl}}.}

\item{table_type}{Either \code{capabilities} or \code{threats}}
}
\value{
Extracted table as a dataframe
}
\description{
Split a sheet of the survey spreadsheet into capabilities or threats
}
