% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_to_disk}
\alias{summarize_to_disk}
\title{Create all summary files and write to disk}
\usage{
summarize_to_disk(simulation_results, results_dir)
}
\arguments{
\item{simulation_results}{Simulation results dataframe.}

\item{results_dir}{Directory to place simulation files.}
}
\value{
Tibble with paths to the created data files.
}
\description{
This is a wrapper around \code{\link{summarize_scenario}} and
\code{\link{summarize_domains}}, calling both functions and writing the
dataframes to a location on disk.
}
\examples{
\dontrun{
data(mc_simulation_results)
summarize_to_disk(mc_simulation_results, results_dir = tempdir())
}
}
