% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{theme_evaluator}
\alias{theme_evaluator}
\title{Default ggplot theme used by all Evaluator-supplied graphics.}
\usage{
theme_evaluator(base_family = "BentonSansRE")
}
\arguments{
\item{base_family}{Font family}
}
\value{
ggplot theme object
}
\description{
Default ggplot theme used by all Evaluator-supplied graphics.
}
