% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_factor.R
\name{tidyrisk_factor}
\alias{tidyrisk_factor}
\alias{new_tidyrisk_factor}
\title{Construct a tidyrisk_factor object}
\usage{
new_tidyrisk_factor(
  samples = double(),
  factor_label = character(),
  details = list()
)

tidyrisk_factor(samples, factor_label, details = list())
}
\arguments{
\item{samples}{samples}

\item{factor_label}{fl}

\item{details}{details}
}
\description{
Construct a tidyrisk_factor object
}
