% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_factor.R
\name{tidyrisk_factory}
\alias{tidyrisk_factory}
\title{Create a tidyrisk_factor sample function}
\usage{
tidyrisk_factory(factor_label = "TC")
}
\arguments{
\item{factor_label}{abbreviation of the OpenFAIR element}
}
\description{
Create a tidyrisk_factor sample function
}
