% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_scenarios}
\alias{validate_scenarios}
\title{Validate scenario data.}
\usage{
validate_scenarios(scenarios, capabilities, domains, mappings)
}
\arguments{
\item{scenarios}{Dataframe of scenarios}

\item{capabilities}{Dataframe of capabilities}

\item{domains}{Dataframe of domain mappings}

\item{mappings}{Dataframe of qualitative mappings}
}
\value{
Invisible NULL.
}
\description{
Validate scenario data.
}
