% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_scenario.R
\name{validate_tidyrisk_scenario}
\alias{validate_tidyrisk_scenario}
\title{Validates that a scenario object is well formed}
\usage{
validate_tidyrisk_scenario(x)
}
\arguments{
\item{x}{An object}
}
\description{
Validates that a scenario object is well formed
}
