% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_factor.R
\name{vec_cast.tidyrisk_factor}
\alias{vec_cast.tidyrisk_factor}
\title{Cast a \code{tidyrisk_factor} vector to a specified type}
\usage{
\method{vec_cast}{tidyrisk_factor}(x, to)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}
}
\description{
Cast a \code{tidyrisk_factor} vector to a specified type
}
