% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_scenario.R
\name{vec_ptype_abbr.tidyrisk_scenario}
\alias{vec_ptype_abbr.tidyrisk_scenario}
\title{Set an abbreviation when displaying an S3 column in a tibble}
\usage{
vec_ptype_abbr.tidyrisk_scenario(x)
}
\arguments{
\item{x}{An object}
}
\description{
Set an abbreviation when displaying an S3 column in a tibble
}
