% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_max_losses}
\alias{calculate_max_losses}
\title{Calculate maximum losses with and without outliers}
\usage{
calculate_max_losses(simulation_results, scenario_outliers = NULL)
}
\arguments{
\item{simulation_results}{Simulation results dataframe.}

\item{scenario_outliers}{Optionnal vector of IDs of outlier scenarios.}
}
\value{
Dataframe.
}
\description{
Calculate the biggest single annual loss for each scenario, as well as
the minimum and maxium ALE across all simulations. Calculation both
with and without outliers (if passed) are returned.
}
