% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_max_losses}
\alias{calculate_max_losses}
\title{Calculate maximum losses with and without outliers}
\usage{
calculate_max_losses(simulation_results, scenario_outliers)
}
\arguments{
\item{simulation_results}{Data.}

\item{scenario_outliers}{Vector of scenario_ids which are outliers}
}
\value{
A dataframe.
}
\description{
Calculate maximum losses with and without outliers
}
