% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capabilities}
\alias{capabilities}
\title{Capabilities}
\format{\describe{
\item{capability_id}{unique id of the capability}
\item{domain_id}{domain id to which the capability applies}
\item{capability}{full text summary of the capability}
\item{diff}{qualitative label of control effectiveness}
}}
\usage{
capabilities
}
\description{
A qualitative dataset of sample capabilities and qualitative level of effectiveness.
}
\keyword{datasets}
