% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{create_templates}
\alias{create_templates}
\title{Create a directory structure for risk analysis, pre-populated with templates}
\usage{
create_templates(base_directory)
}
\arguments{
\item{base_directory}{Parent directory under which to create starter files.}
}
\value{
A dataframe of the starter filenames, along with a flag on whether a file was copied.
}
\description{
Copy the sample files into an \code{inputs} subdirectory. This makes the starter
files available for customizing and data collection. The \code{inputs}
directory will be created if not already present. Preexisting files, if
present, will not be overwritten. Also creates an empty \code{results}
subdirectory as a default location for evaluator output.
}
\examples{
\dontrun{
create_templates("~/evaluator")
}
}
