% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{derive_control_key}
\alias{derive_control_key}
\title{Derive control difficulty parameters for a given qualitative scenario}
\usage{
derive_control_key(capability_ids, capabilities)
}
\arguments{
\item{capability_ids}{Comma-delimited list of capabilities in scope for a scenario.}

\item{capabilities}{Dataframe of master list of all qualitative capabilities.}

\item{mappings}{Qualitative mappings dataframe.}
}
\value{
A named list of quantitative estimate parameters for the capabilities
applicable to a given scenario.
}
\description{
Given a comma-separated list of control IDs in a scenario, identify
the qualitative rankings associated with each scenario, convert to
their quantitative parameters, and return a dataframe of the set of
parameters.
}
\examples{
data(capabilities)
capability_ids <- c("1, 3")
derive_control_key(capability_ids, capabilities)
}
