% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{derive_controls}
\alias{derive_controls}
\title{Derive control difficulty parameters for a given qualitative scenario}
\usage{
derive_controls(labels, capabilities, mappings, id = "None")
}
\arguments{
\item{labels}{Comma delimited list of qualitative labels.}

\item{capabilities}{Character string of the type (tef, tc, diff, etc.).}

\item{mappings}{Qualitative mappings dataframe.}

\item{id}{Character string of the type (tef, tc, diff, etc.).}
}
\value{
List-wrapped dataframe of estimate parameters
}
\description{
Given a comma separated list of control IDs in a scenario, identify
the qualitative rankings associated with each scenario, convert to
their quantitative parameters, and return a dataframe of the set of
parameters.
}
