% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dollar_millions}
\alias{dollar_millions}
\title{Format dollar amounts in terms of millions of USD}
\usage{
dollar_millions(x)
}
\arguments{
\item{x}{A number.}
}
\value{
String in the format of $xM.
}
\description{
Given a number, return a string formatted in terms of millions of dollars.
}
\examples{
dollar_millions(1.523 * 10^6)
}
