% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain_summary}
\alias{domain_summary}
\title{Domain-level risk summary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 2 columns.}
\usage{
domain_summary
}
\description{
A dataset of quantified information security risk, summarized at the
domain level.
}
\details{
\describe{
\item{domain_id}{abbreviated name of the domain}
\item{domain}{full title of the domain}
\item{ale}{annual loss expected, in US dollars}
\item{simulation}{simulation id number}
}
}
\keyword{datasets}
