% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domain_summary}
\alias{domain_summary}
\title{Domain-level risk summary}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 140000 rows and 4 columns.}
\usage{
domain_summary
}
\description{
A dataset of quantified information security risk, summarized at the
domain level.
}
\details{
\describe{
  \item{ale}{annual loss expected, in US dollars}
  \item{domain}{full title of the domain}
  \item{domain_id}{abbreviated name of the domain}
  \item{simulation}{simulation id number}
}
}
\keyword{datasets}
