% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domains}
\alias{domains}
\title{Domain mappings}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 2 columns.}
\usage{
domains
}
\description{
A dataset of domains and domain IDs.
}
\details{
\describe{
\item{domain_id}{abbreviated name of the domain}
\item{domain}{full title of the domain}
}
}
\keyword{datasets}
