% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_scenarios}
\alias{encode_scenarios}
\title{Encode qualitative data to quantitative parameters.}
\usage{
encode_scenarios(scenarios, capabilities, mappings)
}
\arguments{
\item{scenarios}{Qualitative risk scenarios dataframe}

\item{capabilities}{Qualitative program capabilities dataframe}

\item{mappings}{Number of simulations for each scenario}
}
\value{
Dataframe of encoded quantitative scenarios
}
\description{
Encode qualitative data to quantitative parameters.
}
