% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\docType{package}
\name{evaluator}
\alias{evaluator}
\alias{evaluator-package}
\title{\code{evaluator} package}
\description{
Quantified Information Risk Simulation Toolkit
}
\details{
See the online documentation located at
\href{https://evaluator.tidyrisk.org/}{https://evaluator.tidyrisk.org/}
}
