% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\docType{package}
\name{evaluator}
\alias{evaluator}
\alias{evaluator-package}
\title{\code{evaluator} package}
\description{
Information Security Risk Simulation Toolkit
}
\details{
See the README on
\href{https://cran.r-project.org/package=googlesheets/README.html}{CRAN}
or \href{https://github.com/davidski/evaluator}{GitHub}
}
