% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{explore_scenarios}
\alias{explore_scenarios}
\title{Launch the Scenario Explorer web application}
\usage{
explore_scenarios(input_directory = "data", results_directory = "results")
}
\arguments{
\item{input_directory}{Location of input files.}

\item{results_directory}{Location of simulation results.}
}
\value{
Invisible NULL.
}
\description{
Evaluator provides a simple Shiny web application for interactive
exploration of simulation results. This allows a user to get an overview
of simluation output without generating an extensive report.
}
\examples{
\dontrun{
explore_scenarios("~/inputs", "~/results")
}
}
