% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{generate_event_outcomes_plot}
\alias{generate_event_outcomes_plot}
\title{Display the distribution of threat events contained vs. realized across
all domains}
\usage{
generate_event_outcomes_plot(domain_summary)
}
\arguments{
\item{domain_summary}{Domain-level summary from \code{domain_summary}.}
}
\value{
ggplot object.
}
\description{
Creates a barbell plot showing the number and percentage of events
contained (not resulting in loss) vs the number and percentage of
loss events (threat events resulting in losses).
}
\examples{
data(domain_summary)
generate_event_outcomes_plot(domain_summary)
}
