% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{generate_event_outcomes_plot}
\alias{generate_event_outcomes_plot}
\title{Display the distribution of threat events contained vs. realized across
all domains}
\usage{
generate_event_outcomes_plot(control_weakness)
}
\arguments{
\item{control_weakness}{Domain-level control weakness from \code{calculate_weak_domains}}
}
\value{
ggplot object
}
\description{
Creates a barbell plot showing the number and percentage of events
contained (not resulting in loss) vs the number and percentage of
loss events (threat events resulting in losses).
}
\examples{
data(simulation_results)
data(domains)
dat <- calculate_weak_domains(simulation_results, domains)
generate_event_outcomes_plot(dat)
}
