% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{Display a heatmap of impact by domain}
\usage{
generate_heatmap(domain_summary)
}
\arguments{
\item{domain_summary}{Simulations summarized at a domain level via \code{summarize_domains}.}
}
\value{
A ggplot object.
}
\description{
Given a domain_summary and a list of all domains, generate a heatmap colored
by the 95\% VaR. This plot displays the domains in which aggregreate risk is
greater than others.
}
\examples{
data(mc_domain_summary)
generate_heatmap(mc_domain_summary)
}
\seealso{
Other result graphs: 
\code{\link{exposure_histogram}()},
\code{\link{generate_event_outcomes_plot}()},
\code{\link{generate_scatterplot-deprecated}},
\code{\link{loss_exceedance_curve}()},
\code{\link{loss_scatterplot}()}
}
\concept{result graphs}
