% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\title{Generate sample analysis report.}
\usage{
generate_report(input_directory = file.path(getwd(), "data"),
  results_directory = file.path(getwd(), "results"),
  output_file = "risk_report.html", ...)
}
\arguments{
\item{input_directory}{Location of input files}

\item{results_directory}{Location of simulation results}

\item{output_file}{Name of the output file to generate.}

\item{...}{Any other parameters to pass straight to \code{rmarkdown::render}}
}
\value{
Default return values of the \code{rmarkdown::render} function.
}
\description{
Generate sample analysis report.
}
