% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{generate_scatterplot}
\alias{generate_scatterplot}
\title{Display a scatterplot for a particular scenario ID.}
\usage{
generate_scatterplot(simulation_results, scenario_id)
}
\arguments{
\item{simulation_results}{Simulation results}

\item{scenario_id}{ID of the scenario to display}
}
\value{
ggplot object
}
\description{
Display a scatterplot for a particular scenario ID.
}
