% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{get_base_fontfamily}
\alias{get_base_fontfamily}
\title{Select a base graphics font family}
\usage{
get_base_fontfamily()
}
\value{
String of the preferred base font.
}
\description{
The Arial Narrow font is preferred. If not available, use a default \code{sans}
family font.
}
\examples{
get_base_fontfamily()
}
