% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{get_base_fontfamily}
\alias{get_base_fontfamily}
\title{Determine fonts we can/should use.}
\usage{
get_base_fontfamily()
}
\value{
String of the preferred base font
}
\description{
Determine fonts we can/should use.
}
