% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_graphs.R
\name{get_base_fontfamily}
\alias{get_base_fontfamily}
\title{Select a base graphics font family}
\usage{
get_base_fontfamily()
}
\value{
String of the preferred base font.
}
\description{
The Benton Sans Regular font is preferred with a fallback of Arial Narrow.
If neither font is available, use a default \code{sans} family font.
}
\examples{
get_base_fontfamily()
}
