% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_capabilities}
\alias{import_capabilities}
\title{Import capabilities from survey spreadsheet}
\usage{
import_capabilities(survey_file = system.file("survey", "survey.xlsx", package
  = "evaluator"), domains)
}
\arguments{
\item{survey_file}{Path to survey XLSX file. Defaults to a sample file if not supplied.}

\item{domains}{Dataframe of domains and domain IDs.}
}
\value{
Extracted capabilities as a dataframe.
}
\description{
Import capabilities from survey spreadsheet
}
