% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_spreadsheet}
\alias{import_spreadsheet}
\title{Import the scenario spreadsheet.}
\usage{
import_spreadsheet(survey_file = system.file("survey", "survey.xlsx", package
  = "evaluator"), domains = NULL, output_dir = "~/data")
}
\arguments{
\item{survey_file}{Path to survey XLSX file. Defaults to a sample file if not supplied.}

\item{domains}{Dataframe of domains and domain IDs. Defaults to built-in sample \code{domains} dataset.}

\item{output_dir}{Output file directory. Defaults to a \code{data}
subdirectory in the current working direction.}
}
\value{
Dataframe of generated files (capabilities.csv and scenarios.csv)
}
\description{
This is a wrapper function around \code{import_scenarios} and
\code{import_capabilities}, calling both functions and writing the dataframes
to a location on disk.
}
