% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_spreadsheet}
\alias{import_spreadsheet}
\title{Import the scenario spreadsheet}
\usage{
import_spreadsheet(survey_file = system.file("survey", "survey.xlsx", package
  = "evaluator"), domains = NULL, output_dir = "~/evaluator/results")
}
\arguments{
\item{survey_file}{Path to survey XLSX file. Defaults to an Evaluator-provided sample spreadsheet.}

\item{domains}{Dataframe of domains and domain IDs. Defaults to built-in sample \code{domains} dataset.}

\item{output_dir}{Output file directory. Defaults to a \code{data}
subdirectory in the current working directory.}
}
\value{
Dataframe of generated files (capabilities.csv and scenarios.csv)
}
\description{
This is a wrapper function around \code{\link{import_scenarios}} and
\code{\link{import_capabilities}}. When invoked, the output of both functions
are written to disk.
}
