% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load input and results files}
\usage{
load_data(input_directory, results_directory)
}
\arguments{
\item{input_directory}{Location of input files}

\item{results_directory}{Location of simulation results}
}
\value{
List of data objects
}
\description{
Load input and results files
}
