% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrisk_scenario.R
\name{new_tidyrisk_scenario}
\alias{new_tidyrisk_scenario}
\alias{tidyrisk_scenario}
\title{Construct a quantitative scenario object}
\usage{
new_tidyrisk_scenario(..., model = "openfair_tef_tc_diff_lm")

tidyrisk_scenario(..., model = "openfair_tef_tc_diff_lm")
}
\arguments{
\item{...}{One or more named OpenFAIR factor with parameters for sampling}

\item{model}{Name of model to run}
}
\description{
Supply one or more named lists in the format of \code{foo_params},
where each \code{foo} is an OpenFAIR factor name (e.g. tef, tc, diff, lm).
Each factor should include a function name (\code{func}) to which the
other named elements in the list are passed as parameters when
sampling.
}
