% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{openfair_example}
\alias{openfair_example}
\title{Launch OpenFAIR demonstration web application.}
\usage{
openfair_example()
}
\value{
Invisible NULL
}
\description{
Launch OpenFAIR demonstration web application.
}
