% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{openfair_example}
\alias{openfair_example}
\title{Launch OpenFAIR demonstration web application}
\usage{
openfair_example(intermediates_dir = tempdir())
}
\arguments{
\item{intermediates_dir}{Location for intermediate knit files.}
}
\value{
Invisible NULL
}
\description{
A simple web application to demonstrate OpenFAIR modeling. This application
allows a user to enter beta PERT parameters and run simulations to see the
distribution of results, with high level summary statistics. As a demonstration
application, only TEF+TC+DIFF+LM parameters may be entered.
}
\examples{
\dontrun{
openfair_example()
}
}
