% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qualitative_scenarios}
\alias{qualitative_scenarios}
\title{Qualitatitve information security risk scenarios}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56 rows and 8 columns.}
\usage{
qualitative_scenarios
}
\description{
A dataset of qualified information security risk scenarios.
}
\details{
\describe{
\item{scenario_id}{id of the scenario, primary key}
\item{scenario}{full text description of the risk scenario}
\item{tcomm}{full text name of threat community}
\item{tef}{qualitative label of the frequency of threat events}
\item{tc}{qualitative label of the threat capability}
\item{lm}{qualitative label of the single loss magnitude}
\item{domain_id}{domain id to which the scenario applies}
\item{controls}{comma seperate string of the controls for the scenario}
}
}
\keyword{datasets}
