% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quantitative_scenarios}
\alias{quantitative_scenarios}
\title{Quantified information risk scenarios}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56 rows and 9 columns.}
\usage{
quantitative_scenarios
}
\description{
A dataset of quantified risk scenarios, with parameters
describing the distribution of each input.
}
\details{
\describe{
\item{scenario}{full text description of the risk scenario}
\item{scenario_id}{id of the scenario, primary key}
\item{tcomm}{full text name of threat community}
\item{domain_id}{domain abbreviation}
\item{controls}{comma separated list of control ids that apply to this scenario}
\item{diff_params}{nested dataframe of the controls and difficulty parameters associated with the scenario}
\item{tef_params}{list of the threat expected frequency parameters}
\item{tc_params}{list of the threat capability parameters}
\item{lm_params}{list of the loss magnitude parameters}
}
}
\keyword{datasets}
