% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quantitative_scenarios}
\alias{quantitative_scenarios}
\title{Quantified information risk scenarios}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56 rows and 6 columns.}
\usage{
quantitative_scenarios
}
\description{
A dataset of quantified risk scenarios, with parameters
describing the distribution of each input.
}
\details{
\describe{
\item{scenario_id}{id of the scenario, primary key}
\item{scenario_description}{full text description of the risk scenario}
\item{tcomm}{description of the threat community}
\item{domain_id}{domain id}
\item{control_descriptons}{named list of the text description of controls involved}
\item{scenario}{\code{\link{evaluator_scen}} objects}
}
}
\keyword{datasets}
