% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_simulations}
\alias{run_simulations}
\title{Run simulations for a scenario}
\usage{
run_simulations(scenario, iterations = 10000L, ale_maximum = NULL,
  verbose = FALSE, simulation_count = NULL)
}
\arguments{
\item{scenario}{A \link{tidyrisk_scenario} object.}

\item{iterations}{Number of iterations to run on each scenario.}

\item{ale_maximum}{Maximum practical annual losses.}

\item{verbose}{Whether verbose console output is requested.}

\item{simulation_count}{\strong{DEPRECATED} Number of simulations to perform.}
}
\value{
Dataframe of raw results.
}
\description{
Given a quantitative scenario object, run an OpenFAIR Monte Carlo simulation
returning a dataframe of results.
}
\examples{
data(quantitative_scenarios)
# run a single scenario through a trivial number (10) of trials
run_simulations(quantitative_scenarios[[1, "scenario"]], 10)
}
