% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{run_simulations}
\alias{run_simulations}
\title{Run simulations for all scenarios.}
\usage{
run_simulations(scenario, model = "openfair_tef_tc_diff_lm",
  simulation_count = 10000L)
}
\arguments{
\item{scenario}{Quantitative scenarios}

\item{model}{OpenFAIR model to use}

\item{simulation_count}{Number of simulations for each scenario}
}
\value{
Dataframe of raw results
}
\description{
Run simulations for all scenarios.
}
