% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_lef}
\alias{sample_lef}
\title{Sample loss event frequency}
\usage{
sample_lef(n, .func = NULL, params = NULL)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{.func}{Function to use to simulate LEF, defaults to \code{\link[stats]{rnorm}}.}

\item{params}{Optional parameters to pass to \code{.func}.}
}
\value{
List containing type ("lef"), samples (as a vector), and details (as a list).
}
\description{
Sample loss event frequency
}
\seealso{
Other OpenFAIR helpers: \code{\link{compare_tef_vuln}},
  \code{\link{get_mean_control_strength}},
  \code{\link{openfair_tef_tc_diff_lm}},
  \code{\link{sample_diff}}, \code{\link{sample_lm}},
  \code{\link{sample_tc}}, \code{\link{sample_vuln}},
  \code{\link{select_loss_opportunities}}
}
\concept{OpenFAIR helpers}
