% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_lm}
\alias{sample_lm}
\title{Given a number of loss events and a loss distribution, calculate losses}
\usage{
sample_lm(func = NULL, params = NULL)
}
\arguments{
\item{func}{Function to use to simulate TEF, defaults to mc2d::rpert.}

\item{params}{Optional parameters to pass to `func`}
}
\value{
List containing type ("lm"), samples (as a vector), and details (as a list).
}
\description{
Given a number of loss events and a loss distribution, calculate losses
}
