% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_tef}
\alias{sample_tef}
\title{Calculate the number of simulated threat event frequencies (TEF)}
\usage{
sample_tef(n = 1, func = NULL, params = NULL)
}
\arguments{
\item{n}{Number of periods to simulate. Defaults to 1.}

\item{func}{Function to use to simulate TEF, defaults to mc2d::rpert}

\item{params}{Optional parameters to pass to `func`}
}
\value{
List containing type ("tef"), samples (as a list), and details (as a list).
}
\description{
Calculate the number of simulated threat event frequencies (TEF)
}
