% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_tef}
\alias{sample_tef}
\title{Calculate the number of simulated threat event frequencies (TEF)}
\usage{
sample_tef(func = NULL, params = NULL)
}
\arguments{
\item{func}{Function to use to simulate TEF, defaults to mc2d::rpert.}

\item{params}{Optional parameters to pass to `func`.}
}
\value{
List containing type ("tef"), samples (as a vector), and details (as a list).
}
\description{
Calculate the number of simulated threat event frequencies (TEF)
}
