% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_tef}
\alias{sample_tef}
\title{Calculate the number of simulated threat event frequencies (TEF)}
\usage{
sample_tef(n, params = NULL, .func = NULL)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{params}{Optional parameters to pass to \code{.func}.}

\item{.func}{Function to use to simulate TEF, defaults to \code{\link[mc2d]{rpert}}.}
}
\value{
List containing type ("tef"), samples (as a vector), and details (as a list).
}
\description{
Calculate the number of simulated threat event frequencies (TEF)
}
\seealso{
Other OpenFAIR models: 
\code{\link{openfair_tef_tc_diff_plm_sr}()}
}
\concept{OpenFAIR models}
