% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_vuln}
\alias{sample_vuln}
\title{Calculate the vulnerability}
\usage{
sample_vuln(func = NULL, params = NULL)
}
\arguments{
\item{func}{Function to use to simulate VULN, defaults to \code{\link[stats]{rbinom}}.}

\item{params}{Optional parameters to pass to \code{func}.}
}
\value{
List containing type ("vuln"), samples (as a vector), and details (as a list).
}
\description{
Calculate the vulnerability
}
\seealso{
Other OpenFAIR helpers: \code{\link{compare_tef_vuln}},
  \code{\link{get_mean_control_strength}},
  \code{\link{openfair_tef_tc_diff_lm}},
  \code{\link{sample_diff}}, \code{\link{sample_lef}},
  \code{\link{sample_lm}}, \code{\link{sample_tc}},
  \code{\link{sample_tef}},
  \code{\link{select_loss_opportunities}}
}
