% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfair.R
\name{sample_vuln}
\alias{sample_vuln}
\title{Calculate the vulnerability}
\usage{
sample_vuln(func = NULL, params = NULL)
}
\arguments{
\item{func}{Function to use to simulate DIFF, defaults to `rbinom``}

\item{params}{Optional parameters to pass to `func`}
}
\value{
List containing type ("diff"), samples (as a vector), and details (as a list).
}
\description{
Calculate the vulnerability
}
