% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_results}
\alias{simulation_results}
\title{Information security risk simulation results}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 560000 rows and 13 columns.}
\usage{
simulation_results
}
\description{
A dataset containing the full results of sample Monte Carlo
simulations of information security risk scenarios.
}
\details{
\describe{
\item{domain_id}{domain abbreviation}
\item{scenario_id}{id of the scenario}
\item{simulation}{id of the simulation}
\item{threat_events}{number of threat events}
\item{loss_events}{number of loss events occurring in the simulation}
\item{vuln}{percentage of threat events that result in losses}
\item{mean_tc_exceedance}{mean amount of TC > DIFF}
\item{mean_diff_exceedance}{mean amount of DIFF > TC}
\item{ale}{annual loss expectancy}
\item{sle_min}{single loss expectancy - minimum}
\item{sle_mean}{single loss expectancy - mean}
\item{sle_median}{single loss expectancy - mean}
\item{sle_max}{single loss expectancy - maximum}
}
}
\keyword{datasets}
