% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{split_sheet}
\alias{split_sheet}
\title{Split a sheet of the survey spreadsheet into either capabilities or threats}
\usage{
split_sheet(dat, table_type = "capabilities")
}
\arguments{
\item{dat}{Raw sheet input from \code{\link[readxl]{read_excel}}.}

\item{table_type}{Either \code{capabilities} or \code{threats}}
}
\value{
Extracted table as a dataframe
}
\description{
The default data collection Excel spreadsheet solicits threat
scenarios and applicable controls for each domain. This function
takes a single sheet from the spreadsheet, as read by \code{\link[readxl]{read_excel}}
and pulls out either the capabilities or threats, as directed by the
user.
}
