% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_domains}
\alias{summarize_domains}
\title{Create domain-level summary of simulation results}
\usage{
summarize_domains(simulation_results, domains)
}
\arguments{
\item{simulation_results}{Simulation results dataframe}

\item{domains}{Domain mappings}
}
\value{
Simulation results summarized by domain
}
\description{
Rolls up simulation results to the domain level by summing ALE across all
scenarios at a domain level (per simulation).
}
\examples{
data(simulation_results)
data(domains)
summarize_domains(simulation_results, domains)
}
