% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_domains}
\alias{summarize_domains}
\title{Create domain level summary of simulation results.}
\usage{
summarize_domains(simulation_results, domains)
}
\arguments{
\item{simulation_results}{Input directory}

\item{domains}{Domain mappings}
}
\value{
Simulation results summarized by domain
}
\description{
Create domain level summary of simulation results.
}
