% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_scenarios}
\alias{summarize_scenarios}
\title{Create scenario level summary of simulation results.}
\usage{
summarize_scenarios(simulation_results)
}
\arguments{
\item{simulation_results}{Simulation results dataframe}
}
\value{
Simulation results summarized by scenario
}
\description{
Create scenario level summary of simulation results.
}
