% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_scenarios}
\alias{validate_scenarios}
\title{Validate scenario data}
\usage{
validate_scenarios(scenarios, capabilities, domains, mappings)
}
\arguments{
\item{scenarios}{Dataframe of scenarios.}

\item{capabilities}{Dataframe of capabilities.}

\item{domains}{Dataframe of domain mappings.}

\item{mappings}{Dataframe of qualitative mappings.}
}
\value{
Invisible NULL.
}
\description{
Run a set of basic consistency checks on the key data inputs (scenarios,
capabilities, domains, and mappings).
}
\details{
Checks for:
- All scenarios IDs are consecutive
- All scenarios are distinct
- There are no gaps in control IDs
- All controls referenced in scenarios are defined in the controls table
- All controls are distinct
}
